rename HeaderBar to AppHeaderBar
