+++ packages/components/src/AppHeaderBar/AppHeaderBar.component.js
@@ -1,215 +1,189 @@
 import React from 'react';
-import {
-	Button,
-	FormGroup,
-	FormControl,
-	MenuItem,
-	Navbar,
-	Nav,
-	NavDropdown,
-	NavItem,
-} from 'react-bootstrap';
+import classNames from 'classnames';
 
-import Icon from '../Icon';
+import Action from '../Actions/Action';
+import ActionDropdown from '../Actions/ActionDropdown';
 import Typeahead from '../Typeahead';
 import theme from './AppHeaderBar.scss';
 
-const NAV_ITEM = 'navItem';
-const DROPDOWN = 'dropdown';
-
-export function renderNavItem(props, index) {
-	const { icon, ...rest } = props;
+function Logo({ id, isFull, onClick }) {
+	const icon = isFull ? 'talend-logo' : 'talend-logo-square';
+	const itemClassName = classNames(['tc-header-bar-action', !isFull && 'separated']);
+	const actionClassName = classNames(['tc-header-bar-logo', isFull && 'full']);
 	return (
-		<NavItem key={index} {...rest}>
-			<Icon name={icon} />
-		</NavItem>
-	);
-}
-
-renderNavItem.propTypes = {
-	icon: React.PropTypes.string,
-	...NavItem.propTypes,
-};
-
-export function renderDropdownItem(props, index) {
-	const { icon, name, ...rest } = props;
-	return (
-		<MenuItem key={index} {...rest}>
-			<Icon name={icon} />
-			{name}
-		</MenuItem>
-	);
-}
-
-renderDropdownItem.propTypes = {
-	icon: React.PropTypes.string,
-	name: React.PropTypes.string,
-	onClick: React.PropTypes.func,
-};
-
-export function renderDropdown(props, index) {
-	let dropdownProps = props.dropdown;
-	if (dropdownProps.onSelect) {
-		dropdownProps = {
-			...props.dropdown,
-			onSelect: (eventKey, event) => dropdownProps.onSelect(event, eventKey),
-		};
-	}
-	return (
-		<NavDropdown {...dropdownProps} key={index}>
-			{props.items.map(renderDropdownItem)}
-		</NavDropdown>
-	);
-}
-renderDropdown.propTypes = {
-	dropdown: React.PropTypes.shape(NavDropdown.propTypes),
-	items: React.PropTypes.arrayOf(
-		React.PropTypes.shape(renderDropdownItem.propTypes),
-	),
-};
-
-export function renderNav(props) {
-	return (
-		<Nav {...props.nav}>
-			{props.navItems.map((itemDef, index) => {
-				const { type, item } = itemDef;
-				switch (type) {
-				case NAV_ITEM:
-					return renderNavItem(item, index);
-				case DROPDOWN:
-					return renderDropdown(item, index);
-				default:
-					return null;
-				}
-			})}
-		</Nav>
-	);
-}
-renderNav.propTypes = {
-	nav: React.PropTypes.shape(Nav.propTypes),
-	navItems: React.PropTypes.arrayOf(
-		React.PropTypes.shape({
-			type: React.PropTypes.oneOf([NAV_ITEM, DROPDOWN]),
-			item: React.PropTypes.oneOfType([
-				React.PropTypes.shape(renderNavItem.propTypes),
-				React.PropTypes.shape(renderDropdown.propTypes),
-			]),
-		}),
-	),
-};
-
-export function renderFormGroup(props, index) {
-	return (
-		<FormGroup key={index} {...props.formgroup}>
-			<FormControl {...props.formcontrol} />
-		</FormGroup>
-	);
-}
-renderFormGroup.propTypes = {
-	formgroup: React.PropTypes.shape(FormGroup.propTypes),
-	formcontrol: React.PropTypes.shape(FormControl.propTypes),
-};
-
-export function renderForm(props, index) {
-	return (
-		<Navbar.Form {...props.form} key={index}>
-			{props.formgroups ? props.formgroups.map(renderFormGroup) : null}
-			<Button {...props.button}>
-				{props.icon ? (<Icon name={props.icon} />) : null}
-				{props.buttonLabel}
-			</Button>
-		</Navbar.Form>
-	);
-}
-renderForm.propTypes = {
-	form: React.PropTypes.shape(Navbar.Form.propTypes),
-	formgroups: React.PropTypes.arrayOf(
-		React.PropTypes.shape(renderFormGroup.propTypes),
-	),
-	button: React.PropTypes.shape(Button.propTypes),
-	buttonLabel: React.PropTypes.string,
-};
-
-export function renderTypeahead(search, index) {
-	return (
-		<Navbar.Form pullRight role="search" key={index}>
-			<Typeahead {...search} />
-		</Navbar.Form>
-	);
-}
-renderTypeahead.propTypes = React.PropTypes.shape(Typeahead.propTypes);
-
-export function renderContent(props, index) {
-	if (props.navs) {
-		return props.navs.map(renderNav);
-	}
-	if (props.forms) {
-		return props.forms.map(renderForm);
-	}
-	if (props.search) {
-		return renderTypeahead(props.search, index);
-	}
-	return null;
-}
-renderContent.propTypes = {
-	navs: React.PropTypes.arrayOf(
-		React.PropTypes.shape(renderNav.propTypes),
-	),
-	forms: React.PropTypes.arrayOf(
-		React.PropTypes.shape(renderForm.propTypes),
-	),
-	search: React.PropTypes.shape(Typeahead.propTypes),
-};
-
-/**
- * The top bar is the place where the user finds useful information and tools
- * to describe the application
- * the logo, the name of the application, an optionnal global search,
- * few icons for feedback & onboarding and a menu to access logout
- * and profile information.
- * @param {object} props   react props
- */
-function AppHeaderBar(props) {
-	let brandLink;
-	if (props.brandLink) {
-		brandLink = (
-			<Button
+		<li className={itemClassName}>
+			<Action
 				bsStyle="link"
-				role="link"
-				{...props.brandLink}
-			>
-				{props.app}
-			</Button>
-		);
-	} else {
-		brandLink = (
-			<span>{props.app}</span>
-		);
-	}
-	return (
-		<Navbar fluid fixedTop inverse className={`tc-app-header-bar ${theme['tc-app-header-bar']}`}>
-			<Navbar.Header>
-				<Navbar.Brand>
-					{brandLink}
-				</Navbar.Brand>
-				<Navbar.Toggle />
-			</Navbar.Header>
-			<Navbar.Collapse>
-				{props.content && props.content.map(renderContent)}
-			</Navbar.Collapse>
-		</Navbar>
+				className={actionClassName}
+				hideLabel
+				id={id}
+				label="Go to Portal"
+				icon={icon}
+				onClick={onClick}
+				tooltipPlacement="bottom"
+			/>
+		</li>
 	);
 }
+
+Logo.propTypes = {
+	id: React.PropTypes.string,
+	isFull: React.PropTypes.bool,
+	onClick: React.PropTypes.func.isRequired,
+};
+
+function Brand({ id, isSeparated, name, onClick }) {
+	const className = classNames(['tc-header-bar-action', isSeparated && 'separated']);
+	return (
+		<li className={className}>
+			<Action
+				bsStyle="link"
+				className="tc-header-bar-brand"
+				id={id}
+				label={name}
+				onClick={onClick}
+				tooltipPlacement="bottom"
+			/>
+		</li>
+	);
+}
+
+Brand.propTypes = {
+	id: React.PropTypes.string,
+	isSeparated: React.PropTypes.bool,
+	name: React.PropTypes.string.isRequired,
+	onClick: React.PropTypes.func.isRequired,
+};
+
+function Environment(props) {
+	return (
+		<li className="tc-header-bar-action">
+			<ActionDropdown
+				bsStyle="link"
+				icon="talend-burger"
+				id={props.id}
+				items={props.items}
+				label={props.label}
+				tooltipPlacement="bottom"
+			/>
+		</li>
+	);
+}
+
+Environment.propTypes = {
+	id: React.PropTypes.string,
+	items: ActionDropdown.propTypes.items,
+	label: ActionDropdown.propTypes.label,
+};
+
+function Search(props) {
+	return (
+		<li className="tc-header-bar-action separated flex">
+			<Typeahead {...props} />
+		</li>
+	);
+}
+
+Search.propTypes = Typeahead.propTypes;
+
+function Help({ id, onClick }) {
+	return (
+		<li className="tc-header-bar-action">
+			<Action
+				bsStyle="link"
+				icon="talend-question-circle"
+				id={id}
+				label="Help"
+				onClick={onClick}
+				tooltipPlacement="bottom"
+			/>
+		</li>
+	);
+}
+
+Help.propTypes = {
+	id: React.PropTypes.string,
+	onClick: React.PropTypes.func.isRequired,
+};
+
+function User({ id, items, name }) {
+	return (
+		<li className="tc-header-bar-action separated">
+			<ActionDropdown
+				bsStyle="link"
+				icon="talend-user-circle"
+				id={id}
+				items={items}
+				label={name}
+				noCaret
+				tooltipPlacement="bottom"
+			/>
+		</li>
+	);
+}
+
+User.propTypes = {
+	id: React.PropTypes.string,
+	items: ActionDropdown.propTypes.items,
+	name: ActionDropdown.propTypes.label,
+};
+
+function Products({ id, items }) {
+	return (
+		<li className="tc-header-bar-action">
+			<ActionDropdown
+				bsStyle="link"
+				id={id}
+				icon="talend-launcher"
+				items={items}
+				label="Apps"
+				noCaret
+				pullRight
+				tooltipPlacement="bottom"
+			/>
+		</li>
+	);
+}
+
+Products.propTypes = {
+	id: React.PropTypes.string,
+	items: ActionDropdown.propTypes.items,
+};
+
+function AppHeaderBar(props) {
+	return (
+		<nav className={classNames([`tc-header-bar, ${theme['tc-header-bar']}`])} >
+			<ul className="tc-header-bar-actions">
+				<Logo {...props.logo} />
+				<Brand {...props.brand} isSeparated={!!props.env} />
+				{props.env && (<Environment {...props.env} />)}
+			</ul>
+			<ul className="tc-header-bar-actions right">
+				{props.search && (<Search {...props.search} />)}
+				<Help {...props.help} />
+				<User {...props.user} />
+				<Products {...props.products} />
+			</ul>
+		</nav>
+	);
+}
+
+AppHeaderBar.Logo = Logo;
+AppHeaderBar.Brand = Brand;
+AppHeaderBar.Environment = Environment;
+AppHeaderBar.Search = Search;
+AppHeaderBar.Help = Help;
+AppHeaderBar.User = User;
+AppHeaderBar.Products = Products;
 
 AppHeaderBar.propTypes = {
-	app: React.PropTypes.string.isRequired,
-	brandLink: React.PropTypes.shape({
-		onClick: React.PropTypes.func,
-		className: React.PropTypes.string,
-	}),
-	content: React.PropTypes.arrayOf(
-		React.PropTypes.shape(renderContent.propTypes),
-	),
+	brand: React.PropTypes.shape(Brand.propTypes).isRequired,
+	logo: React.PropTypes.shape(Logo.propTypes).isRequired,
+	env: React.PropTypes.shape(Environment.propTypes),
+	search: React.PropTypes.shape(Search.propTypes),
+	help: React.PropTypes.shape(Help.propTypes).isRequired,
+	user: React.PropTypes.shape(User.propTypes).isRequired,
+	products: React.PropTypes.shape(Products.propTypes).isRequired,
 };
 
 export default AppHeaderBar;
