+++ packages/components/src/AppHeaderBar/AppHeaderBar.scss
@@ -1,5 +1,84 @@
-.tc-app-header-bar {
-	:global(.btn):global(.navbar-brand):hover {
-		background-color: transparent;
+$background-color-on-hover: shade($brand-primary, 30);
+
+.tc-header-bar {
+	background-color: $brand-primary-darker;
+	color: $white;
+	display: flex;
+	position: fixed;
+	top: 0;
+	width: 100%;
+	z-index: 1030;
+
+	:global(.tc-header-bar-actions) {
+		align-items: center;
+		display: flex;
+		padding: 0;
+		margin: 0;
+
+		&:global(.right) {
+			margin-left: auto;
+		}
+
+		:global(.tc-header-bar-action) {
+			height: 100%;
+			list-style: none;
+
+			&:global(.separated):after {
+				content: '|';
+			}
+
+			&:global(.flex) {
+				display: flex;
+				align-items: center;
+			}
+		}
+
+		:global(.tc-header-bar-logo):global(.btn):global(.btn-link) {
+			line-height: $svg-lg-size;
+			padding: 12px $padding-normal;
+
+			:global(.tc-svg-icon) {
+				height: $svg-lg-size;
+				vertical-align: top;
+				width: $svg-lg-size;
+			}
+
+			&:global(.full) :global(.tc-svg-icon) {
+				width: 75px;
+			}
+		}
+
+		:global(.tc-header-bar-brand):global(.btn):global(.btn-link) {
+			font-size: $font-size-base;
+			text-transform: uppercase;
+		}
+	}
+
+	:global(.btn):global(.btn-link) {
+		color: $white;
+		font-size: $font-size-small;
+		height: 100%;
+		text-transform: none;
+
+		&:hover,
+		&:active {
+			background-color: $background-color-on-hover;
+		}
+	}
+
+	:global(.dropdown) {
+		height: 100%;
+
+		:global(.caret) {
+			border-top-color: $white;
+			border-right-color: $white;
+			box-shadow: 1px -1px 0 $white;
+		}
+
+		&:global(.open) {
+			:global(.btn):global(.btn-link) {
+				background-color: $background-color-on-hover;
+			}
+		}
 	}
 }
