--- packages/components/src/HeaderBar/HeaderBar.snapshot.test.js
@@ -1,5 +1,0 @@
-import initStoryshots from 'storyshots';
-
-initStoryshots({
-	configPath: 'src/HeaderBar/__snapshots__',
-});
