--- packages/components/src/HeaderBar/index.js
@@ -1,3 +1,0 @@
-import HeaderBar from './HeaderBar.component';
-
-export default HeaderBar;
