+++ packages/components/src/index.js
@@ -78,7 +78,6 @@ import ConfirmDialog from './ConfirmDialog';
 import Dialog from './Dialog';
 import Drawer from './Drawer';
 import Enumeration from './Enumeration';
-import HeaderBar from './HeaderBar';
 import TreeView from './TreeView';
 import Icon from './Icon';
 import IconsProvider from './IconsProvider';
@@ -106,7 +105,6 @@ export {
 	CollapsiblePanel,
 	Dialog,
 	Enumeration,
-	HeaderBar,
 	ConfirmDialog,
 	Drawer,
 	TreeView,
