--- packages/components/stories/HeaderBar.js
@@ -1,194 +1,0 @@
-import React from 'react';
-import { storiesOf, action } from '@kadira/storybook';
-import Immutable from 'immutable';  // eslint-disable-line import/no-extraneous-dependencies
-
-import { HeaderBar, IconsProvider } from '../src';
-
-const typeaheadItems = [
-	{
-		title: 'category 1',
-		icon: {
-			name: 'talend-filter',
-			title: 'icon',
-		},
-		suggestions: [
-			{
-				title: 'le title 1',
-				description: 'description: Uxoresque est in pacto est marito est hastam nomine in eos discessura incredibile tempus ardore.',
-			},
-			{
-				title: 'title 2 les elephants elementaires ont des aile ',
-				description: 'description: Aut aut cum satis inter Epicuri quidem cum erat inquam controversia autem mihi utrumque Attico.',
-			},
-		],
-	},
-	{
-		title: 'category 2',
-		icon: {
-			name: 'talend-star',
-			title: 'icon',
-		},
-		suggestions: [
-			{
-				title: 'title 3',
-				description: 'description: In sanciatur libere audeamus exspectemus amicitia et dum ne audeamus causa monendum honesta studium valeat.',
-			},
-		],
-	},
-	{
-		title: 'category 3',
-		icon: {
-			name: 'talend-share-alt',
-			title: 'icon',
-		},
-		suggestions: [
-			{
-				title: 'title 4',
-				description: 'description: Praesentibus genero ne in Africani mandavi saepius ipsam C in libro et hoc Laeli cum.',
-			},
-			{
-				title: 'title 5',
-				description: 'description: Feceris unde tot illo tot clientes dederis numerando et indiscretus cum paria et unde ubi.',
-			},
-			{
-				title: 'title 6',
-				description: 'description: Gradu quos cedentium sunt appeterent ita ancoralia instar luna sunt etiam ubi incendente nihil observabant.',
-			},
-		],
-	},
-];
-
-const props = {
-	brand: {
-		id: 'header-brand',
-		name: 'Example App Name',
-		onClick: action('onApplicationNameClick'),
-	},
-	logo: {
-		id: 'header-logo',
-		onClick: action('onLogoClick'),
-	},
-	search: {
-		icon: {
-			name: 'talend-search',
-			title: 'icon',
-			bsStyle: 'link',
-		},
-		id: 'header-search',
-		onToggle: action('onSearchClick'),
-	},
-	help: {
-		id: 'header-help',
-		onClick: action('onHelpClick'),
-	},
-	user: {
-		id: 'header-user',
-		items: [
-			{
-				icon: 'talend-cog',
-				label: 'Settings',
-				onClick: action('onSettingsClick'),
-			},
-		],
-		name: 'User NAME',
-	},
-	products: {
-		id: 'header-products',
-		items: [
-			{
-				icon: 'talend-logo-square',
-				label: 'Portal',
-				onClick: action('onPortalClick'),
-			},
-		],
-	},
-};
-
-const decoratedStories = storiesOf('HeaderBar', module)
-	.addDecorator(story => (
-		<div>
-			{story()}
-			<div className="container" style={{ paddingTop: 40 }} />
-			<IconsProvider />
-		</div>
-	));
-
-decoratedStories
-	.addWithInfo('default', () => {
-		const headerProps = Immutable.fromJS(props).toJS();
-		return <HeaderBar {...headerProps} />;
-	})
-	.addWithInfo('with full logo', () => {
-		const headerProps = Immutable.fromJS(props).toJS();
-		headerProps.logo.isFull = true;
-		return <HeaderBar {...headerProps} />;
-	})
-	.addWithInfo('with environment dropdown', () => {
-		const headerProps = Immutable.fromJS(props).toJS();
-		headerProps.env = {
-			id: 'header-environment',
-			items: [
-				{
-					label: 'Runtime Environment',
-					onClick: action('onEnvClick'),
-				},
-			],
-			label: 'Default',
-		};
-		return <HeaderBar {...headerProps} />;
-	})
-	.addWithInfo('with search input', () => {
-		const headerProps = Immutable.fromJS(props).toJS();
-		headerProps.search = {
-			icon: {
-				name: 'talend-search',
-			},
-			placeholder: 'Search...',
-			onBlur: action('onSearchBlur'),
-			onChange: action('onSearchChange'),
-		};
-		return <HeaderBar {...headerProps} />;
-	})
-	.addWithInfo('while searching', () => {
-		const headerProps = Immutable.fromJS(props).toJS();
-		headerProps.search = {
-			icon: {
-				name: 'talend-search',
-			},
-			position: 'right',
-			value: 'le',
-			searching: true,
-			onBlur: action('onSearchBlur'),
-			onChange: action('onSearchChange'),
-		};
-		return <HeaderBar {...headerProps} />;
-	})
-	.addWithInfo('with search results', () => {
-		const headerProps = Immutable.fromJS(props).toJS();
-		headerProps.search = {
-			icon: {
-				name: 'talend-search',
-			},
-			position: 'right',
-			value: 'le',
-			items: typeaheadItems,
-			onBlur: action('onSearchBlur'),
-			onChange: action('onSearchChange'),
-			onSelect: action('onSearchResultSelect'),
-		};
-		return <HeaderBar {...headerProps} />;
-	})
-	.addWithInfo('with no search result', () => {
-		const headerProps = Immutable.fromJS(props).toJS();
-		headerProps.search = {
-			icon: {
-				name: 'talend-search',
-			},
-			position: 'right',
-			value: 'le',
-			items: [],
-			onBlur: action('onSearchBlur'),
-			onChange: action('onSearchChange'),
-		};
-		return <HeaderBar {...headerProps} />;
-	});
